+++ lib/impl/dasher/list.ex
@@ -1,9 +1,11 @@
 defimpl Digger.Dasher, for: List do
   alias Digger.Dasher
+  alias Digger.Opts.Dasher, as: Opts
 
-  def dasherize([]), do: []
+  def dasherize([], _opts), do: []
 
-  def dasherize([h | t]) do
-    [Dasher.dasherize(h)] ++ Dasher.dasherize(t)
+  def dasherize([h | t], opts) do
+    opts = Opts.set_options(opts)
+    [Dasher.dasherize(h, opts)] ++ Dasher.dasherize(t, opts)
   end
 end