+++ lib/protocols/dasher_protocol.ex
@@ -11,6 +11,9 @@ defprotocol Digger.Dasher do
   'Dasherize' a valid Types.data_type according to the protocol implementation
   """
 
-  @spec dasherize(Types.data_type) :: Types.valid_return_type
-  def dasherize(data_type)
+  @spec dasherize(Types.data_type(), keyword()) :: Types.valid_return_type()
+  def dasherize(
+        data_type,
+        opts \\ [type: :key, key_transform: :dasherize, value_transform: :none]
+      )
 end