+++ lib/protocols/lower_caser_protocol.ex
@@ -11,6 +11,9 @@ defprotocol Digger.LowerCaser do
   Lower case first letter of a valid Types.data_type according to the protocol implementation
   """
 
-  @spec lowercase_first(Types.data_type) :: Types.valid_return_type
-  def lowercase_first(data_type)
+  @spec lowercase_first(Types.data_type(), keyword()) :: Types.valid_return_type()
+  def lowercase_first(
+        data_type,
+        opts \\ [type: :key, key_transform: :lower, value_transform: :none]
+      )
 end