+++ lib/protocols/snake_caser_protocol.ex
@@ -9,6 +9,6 @@ defprotocol Digger.SnakeCaser do
   snake_case a valid Types.data_type according to the protocol implementation
   """
 
-  @spec snake_case(Types.data_type) :: Types.valid_return_type
-  def snake_case(data_type)
+  @spec snake_case(Types.data_type(), keyword()) :: Types.valid_return_type()
+  def snake_case(data_type, opts \\ [type: :key, key_transform: :snake, value_transform: :none])
 end